Re-arrange folder structure for the dart side
