+++ lib/notifications.dart
@@ -1,0 +1,1 @@
+export 'src/notification_service.dart';
