+++ lib/src/audio_cache.dart
@@ -4,12 +4,12 @@ import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart' show rootBundle;
-import 'package:path_provider/path_provider.dart';
 import 'package:http/http.dart' as http;
+import 'package:path_provider/path_provider.dart';
 
-import 'audioplayers.dart';
-import 'player_mode.dart';
-import 'release_mode.dart';
+import 'api/player_mode.dart';
+import 'api/release_mode.dart';
+import 'audioplayer.dart';
 
 /// This class represents a cache for Local Assets to be played.
 ///
