+++ lib/src/audioplayers_web.dart
@@ -1,0 +1,192 @@
+import 'dart:async';
+import 'dart:html';
+
+import 'package:flutter/services.dart';
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+
+import 'api/release_mode.dart';
+
+class WrappedPlayer {
+  double? pausedAt;
+  double currentVolume = 1.0;
+  ReleaseMode currentReleaseMode = ReleaseMode.RELEASE;
+  String? currentUrl;
+  bool isPlaying = false;
+
+  AudioElement? player;
+
+  void setUrl(String url) {
+    currentUrl = url;
+
+    stop();
+    recreateNode();
+    if (isPlaying) {
+      resume();
+    }
+  }
+
+  void setVolume(double volume) {
+    currentVolume = volume;
+    player?.volume = volume;
+  }
+
+  void recreateNode() {
+    if (currentUrl == null) {
+      return;
+    }
+    player = AudioElement(currentUrl);
+    player?.loop = shouldLoop();
+    player?.volume = currentVolume;
+  }
+
+  bool shouldLoop() => currentReleaseMode == ReleaseMode.LOOP;
+
+  void setReleaseMode(ReleaseMode releaseMode) {
+    currentReleaseMode = releaseMode;
+    player?.loop = shouldLoop();
+  }
+
+  void release() {
+    _cancel();
+    player = null;
+  }
+
+  void start(double position) {
+    isPlaying = true;
+    if (currentUrl == null) {
+      return; // nothing to play yet
+    }
+    if (player == null) {
+      recreateNode();
+    }
+    player?.play();
+    player?.currentTime = position;
+  }
+
+  void resume() {
+    start(pausedAt ?? 0);
+  }
+
+  void pause() {
+    pausedAt = player?.currentTime as double?;
+    _cancel();
+  }
+
+  void stop() {
+    pausedAt = 0;
+    _cancel();
+  }
+
+  void _cancel() {
+    isPlaying = false;
+    player?.pause();
+    if (currentReleaseMode == ReleaseMode.RELEASE) {
+      player = null;
+    }
+  }
+}
+
+class AudioplayersPlugin {
+  // players by playerId
+  Map<String, WrappedPlayer> players = {};
+
+  static void registerWith(Registrar registrar) {
+    final channel = MethodChannel(
+      'xyz.luan/audioplayers',
+      const StandardMethodCodec(),
+      registrar,
+    );
+
+    final instance = AudioplayersPlugin();
+    channel.setMethodCallHandler(instance.handleMethodCall);
+  }
+
+  WrappedPlayer getOrCreatePlayer(String playerId) {
+    return players.putIfAbsent(playerId, () => WrappedPlayer());
+  }
+
+  Future<WrappedPlayer> setUrl(String playerId, String url) async {
+    final player = getOrCreatePlayer(playerId);
+
+    if (player.currentUrl == url) {
+      return player;
+    }
+
+    player.setUrl(url);
+    return player;
+  }
+
+  ReleaseMode parseReleaseMode(String value) {
+    return ReleaseMode.values.firstWhere((e) => e.toString() == value);
+  }
+
+  Future<dynamic> handleMethodCall(MethodCall call) async {
+    final method = call.method;
+    final args = call.arguments as Map<dynamic, dynamic>;
+    final playerId = args['playerId'] as String;
+    switch (method) {
+      case 'setUrl':
+        {
+          final url = args['url'] as String;
+          await setUrl(playerId, url);
+          return 1;
+        }
+      case 'play':
+        {
+          final url = args['url'] as String;
+
+          // TODO(luan) think about isLocal (is it needed or not)
+
+          final volume = args['volume'] as double? ?? 1.0;
+          final position = args['position'] as double? ?? 0;
+          // web does not care for the `stayAwake` argument
+
+          final player = await setUrl(playerId, url);
+          player.setVolume(volume);
+          player.start(position);
+
+          return 1;
+        }
+      case 'pause':
+        {
+          getOrCreatePlayer(playerId).pause();
+          return 1;
+        }
+      case 'stop':
+        {
+          getOrCreatePlayer(playerId).stop();
+          return 1;
+        }
+      case 'resume':
+        {
+          getOrCreatePlayer(playerId).resume();
+          return 1;
+        }
+      case 'setVolume':
+        {
+          final volume = args['volume'] as double? ?? 1.0;
+          getOrCreatePlayer(playerId).setVolume(volume);
+          return 1;
+        }
+      case 'setReleaseMode':
+        {
+          final releaseMode = parseReleaseMode(args['releaseMode'] as String);
+          getOrCreatePlayer(playerId).setReleaseMode(releaseMode);
+          return 1;
+        }
+      case 'release':
+        {
+          getOrCreatePlayer(playerId).release();
+          return 1;
+        }
+      case 'seek':
+      case 'setPlaybackRate':
+      default:
+        throw PlatformException(
+          code: 'Unimplemented',
+          details:
+              "The audioplayers plugin for web doesn't implement the method '$method'",
+        );
+    }
+  }
+}
