+++ lib/web/audioplayers_web.dart
@@ -1,0 +1,1 @@
+export '../src/audioplayers_web.dart';
